Refactor DJing section; move component to a dedicated folder, update header description, and clean up package.json
