+++ package.json
@@ -6,6 +6,7 @@
 		"react-router-dom": "^7.4.1",
 		"typescript": "^4.0.0"
 	},
+	"description": "A personal website showcasing my CV, interests in crypto, technology, DJing, and travel.",
 	"devDependencies": {
 		"@types/react": "^19.0.12",
 		"@types/react-dom": "^19.0.4",
@@ -22,7 +23,6 @@
 		"tailwindcss": "^3.4.1",
 		"vite": "^4.0.0"
 	},
-	"description": "A personal website showcasing my CV, interests in crypto, technology, DJing, and travel.",
 	"keywords": [
 		"cardacci",
 		"engineer",
@@ -44,5 +44,6 @@
 		"lint": "eslint --ext .js,.jsx,.ts,.tsx src/",
 		"preview": "vite preview"
 	},
+	"type": "module",
 	"version": "1.0.0"
 }
